Disallow adding an item without a description

+++ main.js
@@ -10,18 +10,21 @@ $(document).ready(function(){
 
   function addItem() {
     let action = inputForm.val();
-    let duedate = $('#duedate').val();
-    duedate = (new Date(duedate)).toUTCString().slice(0,-13);
-    
-    let $item = $('<tr>').addClass('item');
-    let $checkbox = $('<input>').attr('type', 'checkbox').click(checkOffItem);
-    let $xMark = $('<a>').attr('href', '#').text('X').addClass('delete').click(deleteItem);
-    $item.append( $('<td>').text(action) )
-         .append( $('<td>').text(duedate) )
-         .append( $('<td>').append($checkbox).append($xMark) );
 
-    $('#list').append($item);
-    inputForm.val('').focus();
+    if (action) {
+      let duedate = $('#duedate').val();
+      duedate = (new Date(duedate)).toUTCString().slice(0,-13);
+      
+      let $item = $('<tr>').addClass('item');
+      let $checkbox = $('<input>').attr('type', 'checkbox').click(checkOffItem);
+      let $xMark = $('<a>').attr('href', '#').text('X').addClass('delete').click(deleteItem);
+      $item.append( $('<td>').text(action) )
+           .append( $('<td>').text(duedate) )
+           .append( $('<td>').append($checkbox).append($xMark) );
+
+      $('#list').append($item);
+      inputForm.val('').focus();
+    }
   }
 
   function deleteItem() {
@@ -38,6 +41,6 @@ $(document).ready(function(){
   }
 
   function handleKey(e) {
-    if (e.charCode === 13 && inputForm.val() !== '') addItem();
+    if (e.charCode === 13) addItem();
   }
 })
